==> _previous/clustering/qc2022/cluster_baseline/02-2_preparation_donnees_style.ipynb <==
# Préparation des variables pour les variables de style

library(tidyverse)
library(factoextra)

# Charger les données
df_pilot1_2022 <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/01_pilot1_2022.rds")

# Charger le fichier des variables
source("01-1_variables.R")

# Préparation du jeu de données pour les variables de style
df_style <- df_pilot1_2022 %>%
  select(all_of(variables_style))

# Boucle pour afficher la table pour chaque variable
for (v in variables_style) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_style[[v]]))
    cat("\n\n")
  }
}

## Transformations

- app_swag

Les variables app_swag sont telles que certaines catégories sont très peu peuplées. Regrouper les styles pour en faire des styles plus généraux.

- noTatoo

Retourner le sens logique pour que vrai soit avec tatou => deviendra Tatoo

- animal

Regrouper certaines classes d'animal pour avoir des catégories plus générales


# Regroupement de app_swag
#-------------------------

df_style <- df_style %>%

  # Regrouper Formel et Chic sous Classique
  mutate(
    app_swag_Classique = if_else(
      app_swag_Formel == 1 | app_swag_Chic == 1,
      1,
      app_swag_Classique
    )
  ) %>%
  select(-app_swag_Formel, -app_swag_Chic) %>%

  # Regrouper HippBoheme, Punk, Rock sous Other
  mutate(
    app_swag_Other = if_else(
      app_swag_HippBoheme == 1 | app_swag_Punk == 1 | app_swag_Rock == 1,
      1,
      app_swag_Other
    )
  ) %>%
  select(-app_swag_HippBoheme, -app_swag_Punk, -app_swag_Rock)

# Changement de sens logique de NoTatoo
#--------------------------------------

df_style <- df_style %>%
  mutate(app_withTattoo = if_else(app_noTattoo == 1, 0, 1)) %>%
  select(-app_noTattoo)

# Regroupement de animal
#-----------------------

df_style <- df_style %>%
  mutate(animal_other = if_else(
    animal_domestic == 1 | animal_farm == 1,
    1,
    0
  )
  ) %>%
  select(-animal_domestic, -animal_farm)

variables_style_clust <- c(
  "id",
  "app_swag_Classique",
  "app_swag_Casual",
  "app_swag_Sport",
  "app_swag_Other",
  "app_withTattoo",
  "animal_cat",
  "animal_dog",
  "animal_other",
  "animal_noPet"
)

# Boucle pour afficher la table pour chaque variable
for (v in variables_style_clust) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_style[[v]]))
    cat("\n\n")
  }
}

# Sauvegarder les données préparées
saveRDS(df_style, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_style.rds")

---

# SUGGESTION : Activité de Clustering

Si souhaité, effectuer un exercice de clustering avec ce sous-groupe de variables pour mieux saisir les données et leur relation. 

Cet exercice est purement exploratoire et sert à voir si les données peuvent être utilisées pour faire du clustering.
