==> _previous/clustering/qc2022/cluster_baseline/02-3_preparation_donnees_sante.ipynb <==
# Préparation des variables de sante

library(tidyverse)
library(factoextra)

# Charger les données
df_pilot1_2022 <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/01_pilot1_2022.rds")

# Charger le fichier des variables
source("01-1_variables.R")

df_sante <- df_pilot1_2022 %>%
  select(all_of(variables_sante))

# Boucle pour afficher la table pour chaque variable
for (v in variables_sante) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_sante[[v]]))
    cat("\n\n")
  }
}

### Transformations

- cons_meat : Devrait être une variable ordinale 
  - 0 Not more than once a month
  - 0.5 Few weekly
  - 1 Daily

- cons_..._Drink : On peut conserver tel quel

- cons_Smoke : Devrait être une variable ordinale
  - 0 Never
  - 0.5 Not daily
  - 1 Daily

# Transformation de cons_meat en variable ordinale
#-------------------------------------------------

df_sante <- df_sante %>%

  mutate(
    cons_Meat = if_else(
      cons_meat_never == 1 | cons_meat_almost_never == 1 | cons_meat_once_month == 1,
      0,
      if_else(
        cons_meat_once_week == 1 | cons_meat_few_week == 1,
        0.5,
        1
      )
    )
  ) %>%
  select(
    -cons_meat_never,
    -cons_meat_almost_never,
    -cons_meat_once_month,
    -cons_meat_once_week,
    -cons_meat_few_week,
    -cons_meat_daily,
    -cons_meat_few_daily
  )

# Transformation de cons_Smoke variable ordinale
#-------------------------------------------------

df_sante <- df_sante %>%

  mutate(
    cons_Smoke = if_else(
      cons_Smoke_never == 1,
      0,
      if_else(
        cons_Smoke_few_times_year == 1 | cons_Smoke_month == 1 | cons_Smoke_once_week == 1 | cons_Smoke_few_times_week == 1,
        0.5,
        1
      )
    )
  ) %>%
  select(
    -cons_Smoke_never,
    -cons_Smoke_few_times_year,
    -cons_Smoke_month,
    -cons_Smoke_once_week,
    -cons_Smoke_few_times_week,
    -cons_Smoke_once_day,
    -cons_Smoke_few_times_day
  )

variables_sante_clust <- c(
  "id",
  "cons_Meat",
  "cons_redWineDrink",
  "cons_whiteWineDrink",
  "cons_roseDrink",
  "cons_spiritDrink",
  "cons_noDrink",
  "cons_Smoke"
)

# Boucle pour afficher la table pour chaque variable
for (v in variables_sante_clust) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_sante[[v]]))
    cat("\n\n")
  }
}

# Sauvegarder les données préparées
saveRDS(df_sante, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_sante.rds")

---

# SUGGESTION : Activité de Clustering

Si souhaité, effectuer un exercice de clustering avec ce sous-groupe de variables pour mieux saisir les données et leur relation. 

Cet exercice est purement exploratoire et sert à voir si les données peuvent être utilisées pour faire du clustering.
