==> _previous/clustering/qc2022/cluster_baseline/02-6_preparation_donnees_ses.ipynb <==
# Préparation des variables de ses

library(tidyverse)
library(factoextra)

# Charger les données
df_pilot1_2022 <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/01_pilot1_2022.rds")

# Charger le fichier des variables
source("01-1_variables.R")

df_ses <- df_pilot1_2022 %>%
  select(all_of(variables_ses))

# Boucle pour afficher la table pour chaque variable
for (v in variables_ses) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_ses[[v]]))
    cat("\n\n")
  }
}

### Transformations

Les variables age, éducation et income devraient être des variables ordinales

# Transformation de age en variable ordinale
#-------------------------------------------

df_ses <- df_ses %>%

  mutate(
    age = if_else(
      age34m == 1,
      0,
      if_else(
        age3554 == 1,
        0.5,
        1
      )
    )
  ) %>%
  select(
    -age34m,
    -age3554,
    -age55p
  )

# Transformation de educ en variable ordinale
#--------------------------------------------

df_ses <- df_ses %>%

  mutate(
    educ = if_else(
      educBHS == 1,
      0,
      if_else(
        educCollege == 1,
        0.5,
        1
      )
    )
  ) %>%
  select(
    -educBHS,
    -educCollege,
    -educUniv
  )

# Transformation de income en variable ordinale
#----------------------------------------------

df_ses <- df_ses %>%

  mutate(
    ses_income_None = 0,
    ses_income_i1to30    = ses_income_i1to30/7,
    ses_income_i31to60   = (1*ses_income_i31to60)/7 + ses_income_i31to60/7,
    ses_income_i61to90   = (2*ses_income_i61to90)/7 + ses_income_i61to90/7,
    ses_income_i91to110  = (3*ses_income_i91to110)/7 + ses_income_i91to110/7,
    ses_income_i111to150 = (4*ses_income_i111to150)/7 + ses_income_i111to150/7,
    ses_income_i151to200 = (5*ses_income_i151to200)/7 + ses_income_i151to200/7,
    ses_income_i201toInf = (ses_income_i201toInf)
  ) %>%
  mutate (
    ses_income = ses_income_None +
      ses_income_i31to60 +
      ses_income_i61to90 +
      ses_income_i91to110 +
      ses_income_i111to150 +
      ses_income_i151to200 +
      ses_income_i201toInf
  ) %>%
  select(
    -ses_income_None,
    -ses_income_i201toInf,
    -ses_income_i31to60,
    -ses_income_i61to90,
    -ses_income_i91to110,
    -ses_income_i111to150,
    -ses_income_i151to200,
    -ses_income_i201toInf
  )

variables_ses_clust <- c(
  "id",
  "male",
  "female",
  "age",
  "langEn",
  "langFr",
  "ses_languageOther",
  "educ",
  "ses_income",
  "immigrant",
  "ses_ethn_White",
  "ses_ethn_Black",
  "ses_ethn_Other",
  "ses_hetero",
  "ses_gai",
  "ses_bisex",
  "ses_sexOri_other"
)

# Boucle pour afficher la table pour chaque variable
for (v in variables_ses_clust) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_ses[[v]]))
    cat("\n\n")
  }
}

# Sauvegarder les données préparées
saveRDS(df_ses, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_ses.rds")

---

# SUGGESTION : Activité de Clustering

Si souhaité, effectuer un exercice de clustering avec ce sous-groupe de variables pour mieux saisir les données et leur relation. 

Cet exercice est purement exploratoire et sert à voir si les données peuvent être utilisées pour faire du clustering.
