==> _previous/clustering/qc2022/cluster_baseline/02-7_preparation_donnees_finales.ipynb <==
# Assemblage des données

Les variables ont été préparées pour le clustering en sous-ensemble de variables.

Assembler toutes les tables de données contenant chaque sous-ensemble de variables.

library(dplyr)

# Assemblage des fichiers

df_pilot1_2022_act         <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_act.rds")
df_pilot1_2022_style       <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_style.rds")
df_pilot1_2022_sante       <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_sante.rds")
df_pilot1_2022_mode_de_vie <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_mode_de_vie.rds")
df_pilot1_2022_commerce    <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_commerce.rds")
df_pilot1_2022_ses         <- readRDS(file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_ses.rds")

df_pilot1_2022_clust <- df_pilot1_2022_act %>%
  left_join(df_pilot1_2022_style, by = "id") %>%
  left_join(df_pilot1_2022_sante, by = "id") %>%
  left_join(df_pilot1_2022_mode_de_vie, by = "id") %>%
  left_join(df_pilot1_2022_commerce, by = "id") %>%
  left_join(df_pilot1_2022_ses, by = "id")

saveRDS(df_pilot1_2022_clust, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022.rds")
